Rename Slider component
